+++ src/SmartCalc/Models/calculation.c
@@ -94,19 +94,7 @@ void DecisionTrigonometricFunction(char number, NodeNumber* *top) {
     }
 }
 
-// void SaveStack() {
-
-// }
-
-// void NodeOutput(NodeNumber* *top) {
-//     NodeNumber* tmp = *top;
-//     for (int i = 0; tmp != NULL; i++) {
-//         printf("STACK:  %Lf\n", tmp->number);
-//         tmp = tmp->next;
-//     }
-// }
-
-long double Calculation(char* output, double x) {
+double Calculation(char* output, double x) {
     char tmp_output[1024] = {0};
     for (int i = 0; output[i] != '\0'; i++) {
         tmp_output[i] = output[i];
